// apps/ui/src/lib/api-client/models/OrgBrandConfig.ts
export interface OrgBrandConfig {
    /**
     * Brand display name
     * @type {string}
     * @memberof OrgBrandConfig
     */
    name?: string;
    /**
     * Brand website URL
     * @type {string}
     * @memberof OrgBrandConfig
     */
    website?: string;
    /**
     * Brand logo URL (SVG or 1024px PNG, transparent background)
     * @type {string | null}
     * @memberof OrgBrandConfig
     */
    logoUrl?: string | null;
    /**
     * Logo license URL
     * @type {string | null}
     * @memberof OrgBrandConfig
     */
    logoLicenseUrl?: string | null;
    /**
     * Alternative brand names
     * @type {Array<string>}
     * @memberof OrgBrandConfig
     */
    aliases?: Array<string>;
    /**
     * @type {OrgBrandConfigCategoryEnum}
     * @memberof OrgBrandConfig
     */
    category?: OrgBrandConfigCategoryEnum;
    /**
     * Patient-facing portal name
     * @type {string | null}
     * @memberof OrgBrandConfig
     */
    portalName?: string | null;
    /**
     * Patient-facing portal URL
     * @type {string | null}
     * @memberof OrgBrandConfig
     */
    portalUrl?: string | null;
    /**
     * Patient-facing portal description (Markdown)
     * @type {string | null}
     * @memberof OrgBrandConfig
     */
    portalDescription?: string | null;
    /**
     * Patient-facing portal logo URL
     * @type {string | null}
     * @memberof OrgBrandConfig
     */
    portalLogoUrl?: string | null;
    /**
     * Portal logo license URL
     * @type {string | null}
     * @memberof OrgBrandConfig
     */
    portalLogoLicenseUrl?: string | null;
    /**
     * Organization city
     * @type {string | null}
     * @memberof OrgBrandConfig
     */
    addressCity?: string | null;
    /**
     * Organization state/province
     * @type {string | null}
     * @memberof OrgBrandConfig
     */
    addressState?: string | null;
    /**
     * Organization postal code
     * @type {string | null}
     * @memberof OrgBrandConfig
     */
    addressPostalCode?: string | null;
    /**
     * Organization country
     * @type {string | null}
     * @memberof OrgBrandConfig
     */
    addressCountry?: string | null;
    /**
     * Brand identifier URI (typically the brand website URL)
     * @type {string}
     * @memberof OrgBrandConfig
     */
    identifier?: string;
}

/**
 * @export
 */
export const OrgBrandConfigCategoryEnum = {
    Prov: 'prov',
    Ins: 'ins',
    Laboratory: 'laboratory',
    Imaging: 'imaging',
    Pharmacy: 'pharmacy',
    HealthInformationNetwork: 'health-information-network',
    HealthDataAggregator: 'health-data-aggregator',
} as const;
export type OrgBrandConfigCategoryEnum = typeof OrgBrandConfigCategoryEnum[keyof typeof OrgBrandConfigCategoryEnum];